+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - 口座開設を複数口座に対応
 - update mongoose
 - update express-validator
+- update @pecorino/domain
 
 ### Deprecated